+++ chat/message/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lg-chat/message
 
+## 5.0.8
+
+### Patch Changes
+
+- @lg-chat/lg-markdown@3.0.8
+
 ## 5.0.7
 
 ### Patch Changes